requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrs==22.1.0
blend-modes==2.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
dlib==19.23.0
fire==0.4.0
flatbuffers==1.12
fvcore==0.1.5.post20220305
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.7.0
idna==3.2
imageio==2.9.0
imutils==0.5.4
iopath==0.1.9
japanize-matplotlib==1.1.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.6.1
libclang==14.0.6
Markdown==3.3.4
matplotlib==3.4.2
mediapipe==0.8.10.1
memory-profiler==0.60.0
networkx==2.6.2
numpy==1.21.0
oauthlib==3.1.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.1
Pillow==8.3.1
portalocker==2.4.0
protobuf==3.17.3
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySimpleGUI==4.60.0
python-dateutil==2.8.2
pytorch-fid==0.2.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.0
setproctitle==1.2.2
six==1.16.0
tabulate==0.8.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2021.7.2
torch==1.10.0+cu113
torch-tb-profiler==0.3.1
torchaudio==0.10.0+cu113
torchgpipe==0.0.7
torchvision==0.11.1+cu113
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.14.1
yacs==0.1.8
